cilium: set output mark for encryption with tunnel enabled case
We currently set the output mark only for subnet enabled cases. But,
we actually want it in tunnel cases as well. And to simplify design
we can enable it in all cases. This will cause the encryption route
table to be used after encryption, which now after previous patch
will use the correct route MTU, for both tunnel and routing modes.

Otherwise encrypted packets could use the main table again which has
route MTU set to encryption overhead + tunnel overhead which results
in dropping (sending icmp MSS exceeded) packets that are close to
the MTU limits before encryption. The result is a smaller MTU than
expected. With the output mark set we instead use the routing table
which will have the high mtu set.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>