bpf: avoid potentially expensive calls into get_hash_recalc
Depending on the NICs hash, we might end up recalculating it
from the kernel flow dissector to get a higher quality L4 hash
on the skb. We used to do this mainly for the skb->hash based
service LB, but nowadays all occurences in the code are for
cilium monitor debug output where it's used as a heuristic to
see that the event messages belong to the same flow. Whether L3
or L4 hash doesn't really matter here, so get rid of these calls.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>