eni/azure: Don't perform mutating operations while instances API is unstable
It is possible for the cloud APIs being used by the operator to get into
a state where POST and PATCH operations are still succeeding while GET
operations are failing. This can result in the operator to continously
creating resources while being unable to ever synchronize the state
successfully. Require a successful synchronization of all resources in
order to continue performing mutating operations.

Signed-off-by: Thomas Graf <thomas@cilium.io>